etl/requirements.txt
web3>=6.5.0
python-dotenv>=1.0.0
eth-utils>=2.2.1
numpy>=1.26.4